+++ git_machete/generated_docs.py
@@ -1447,9 +1447,11 @@ long_docs: Dict[str, str] = {
 
         Walk the branches in the order as they occur in branch layout file.
         By default, `traverse` starts from the current branch.
-        This behavior can, however, be customized using options: `--start-from=`, `--whole` or `-w`, `-W`.
+        This behavior can, however, be customized using options: `--start-from=...`, `--whole` (`-w`) or `-W`.
 
-        For each branch, the command:
+        For each branch, the command performs the following actions:
+
+        <b>Sync to parent</b>
 
            * detects if the branch is merged (<gray>gray</gray> edge) to its parent (aka upstream):
 
@@ -1460,6 +1462,8 @@ long_docs: Dict[str, str] = {
              - asks the user whether to <b>rebase</b> (default) or merge (if `--merge` passed) the branch onto into its upstream branch
               — equivalent to `git machete update`;
 
+        <b>Sync to remote</b>
+
            * if the branch is not tracked on a remote, is ahead of its remote counterpart, or diverged from the counterpart &
              has newer head commit than the counterpart:
 
@@ -1468,10 +1472,14 @@ long_docs: Dict[str, str] = {
            * otherwise, if the branch diverged from the remote counterpart & has older head commit than the counterpart:
 
              - asks the user whether to <b>reset</b> (`git reset --keep`) the branch to its remote counterpart
+              — note that rebase/merge isn't suggested in that case;
 
            * otherwise, if the branch is behind its remote counterpart:
 
-             - asks the user whether to <b>pull</b> the branch;
+             - asks the user whether to <b>pull</b> the branch
+              — note that rebase/merge isn't suggested in that case;
+
+        <b>Sync to GitHub/GitLab</b>
 
            * if `-H`/`--sync-github-prs` or `-L`/`--sync-gitlab-mrs` option is present:
 
@@ -1481,13 +1489,11 @@ long_docs: Dict[str, str] = {
               and its base/target branch in GitHub/GitLab is different than the upstream in machete file
               (just as `git machete github retarget-pr` and `git machete gitlab retarget-mr` would do);
 
-           * and finally, if any of the above operations has been successfully completed:
+        <b>Status</b>
+
+           * finally, if any of the above operations has been successfully completed:
 
              - prints the updated `status`.
-
-        By default `traverse` asks if the branch should be pushed. This behavior can, however, be changed with the `machete.traverse.push` configuration key.
-        It can also be customized using options: `--[no-]push` or `--[no-]push-untracked` — the order of the flags defines their precedence over each other
-        (the one on the right overriding the ones on the left). More on them in the <b>Options</b> section below.
 
         If the traverse flow is stopped (typically due to merge/rebase conflicts), just run `git machete traverse` after the merge/rebase is finished.
         It will pick up the walk from the current branch.